someone might someday find something interesting here.
the main thing I think is interesting is the tweak to emacs org babel
to make it useful for this type of project. being able to hit "^C^C"
inside a SRC block that says just "make nconfig" and have it work
right is fairly awesome.

the unionfs+systemd-nspawn containerization is kinda cool too.